Fix Issue #19
Make the title for the sidebar to position absolute instead since it
will not be changing in height. Keep the Message a Moderator button
floating right with the same width so it will be below the Moderato
title. The content should then stay on the bottom of the button now no
matter the height.