Updated team flairs to 2017 season
Updated all regions except GPL to 2017 season. Teams that didn't
requalify were removed from the flairs, so this may cause issues with
user flairs getting reset and defaulting to TSM, although it appears
you've done this before so not sure if you have a system in place to fix
this or not.
I also have a script to re-order the selector list if you'd like it as I
see the new champs and teams are out of order as of current.