updated sidebar buttons
Updated sidebar button so that the whole button is clicked instead of just a element field. http://i.imgur.com/4SAFCmp.png and here's how it was before http://i.imgur.com/WHAUM9q.png

Also centered text so that it's inline with the rest of the buttons. I don't work with less so i didn't compile this, sorry.